Unify all board implementations
