+++ src/main/java/pl/games/checkers/ai/Board.java
@@ -62,4 +62,21 @@ public abstract class Board<T extends Pawn> implements Copier<Pawn[][]> {
         return this;
     }
 
+    @Override
+    public Pawn[][] copy() {
+        Pawn[][] pawnsCopy = new Pawn[getHeight()][getWidth()];
+
+        for (int row = 0; row < getHeight(); row++) {
+            for (int col = 0; col < getWidth(); col++) {
+                if (getPawn(row, col) != null) {
+                    pawnsCopy[row][col] = getPawn(row, col).copy();
+                } else {
+                    pawnsCopy[row][col] = null;
+                }
+            }
+        }
+
+        return pawnsCopy;
+    }
+
 }
